Renamed some files according to their purpose.
